research/fedhca2/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
scikit-image>=0.20.0
timm>=0.9.0
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
numpy>=1.24.0
Pillow>=9.0.0
matplotlib>=3.6.0

# NVFLARE
nvflare>=2.4.0


